// projects/markets-web-app/src/generated/points-markets-api.ts
export interface components {
  schemas: {
// ...
    AuctionEligibilityRequest: {
      auctionCommandId: components["schemas"]["OpaqueId"];
      auctionCommandHash: components["schemas"]["Sha256Hash"];
      items: components["schemas"]["AuctionEligibilityRequestItem"][];
    };
// ...
    AuctionEligibilityResponse: {
      data: components["schemas"]["AuctionEligibilityData"];
      meta: components["schemas"]["RequestMeta"];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}